Fixed some bugs with the grabbing script.
Fixed a bug where player couldn't pickup the map and compass. It was caused, because I was using a function that stopped working for some reason. Added that objects don't move after leaving them. Added that objects are held where the player grabbed them, not in the middle. Added that the player picks up the nearest object.